==> processing/01a_check_mriqc_completion.ipynb <==
# Evaluate completion of MRIQC on datasets

import os.path as op
from glob import glob

import pandas as pd

# Constants
project_dir = "/home/data/nbc/misc-projects/Salo_PowerReplication/"

dsets = {
    "dset-camcan": [
        "{}_T1w.html",
        "{}_T2w.html",
        "{}_task-movie_echo-1_bold.html",
        "{}_task-movie_echo-2_bold.html",
        "{}_task-movie_echo-3_bold.html",
        "{}_task-movie_echo-4_bold.html",
        "{}_task-movie_echo-5_bold.html",
    ],
    "dset-cambridge": [
        "{}_T1w.html",
        "{}_task-rest_echo-1_bold.html",
        "{}_task-rest_echo-2_bold.html",
        "{}_task-rest_echo-3_bold.html",
        "{}_task-rest_echo-4_bold.html",
    ],
    "dset-dupre": [
        "{}_T1w.html",
        "{}_task-rest_run-01_echo-1_bold.html",
        "{}_task-rest_run-01_echo-2_bold.html",
        "{}_task-rest_run-01_echo-3_bold.html",
    ],
    "dset-dalenberg": [
        "{}_T1w.html",
        "{}_task-images_echo-1_bold.html",
        "{}_task-images_echo-2_bold.html",
        "{}_task-images_echo-3_bold.html",
    ],
    "dset-cohen": [
        "{}_T1w.html",
        "{}_task-bilateralfingertapping_echo-1_bold.html",
        "{}_task-bilateralfingertapping_echo-2_bold.html",
        "{}_task-bilateralfingertapping_echo-3_bold.html",
        "{}_task-bilateralfingertapping_echo-4_bold.html",
    ],
}

for dset, target_files in dsets.items():
    print(dset)
    dset_dir = op.join(project_dir, dset)
    mriqc_dir = op.join(dset_dir, "derivatives/mriqc/")
    if not op.isdir(mriqc_dir):
        print("\tDataset not yet processed.")
        continue
    participants_file = op.join(dset_dir, "participants.tsv")
    participants_df = pd.read_table(participants_file)
    subject_list = participants_df["participant_id"].tolist()
    failed_subjects = []
    for sub in subject_list:
        tfs = [op.join(mriqc_dir, temp.format(sub)) for temp in target_files]
        if not all(op.isfile(tf) for tf in tfs):
            failed_subjects.append(sub)

    if failed_subjects:
        print("\t" + "\n\t".join(failed_subjects))

for dset, target_files in dsets.items():
    print(dset)
    dset_dir = op.join(project_dir, dset)
    mriqc_dir = op.join(dset_dir, "derivatives/mriqc/")
    if not op.isdir(mriqc_dir):
        print("\tDataset not yet processed.")
        continue
    participants_file = op.join(dset_dir, "participants.tsv")
    participants_df = pd.read_table(participants_file)
    subject_list = participants_df["participant_id"].tolist()
    failed_subjects = []
    for sub in subject_list:
        tfs = [op.join(mriqc_dir, temp.format(sub)) for temp in target_files]
        for tf in tfs:
            if not op.isfile(tf):
                print("\t{}".format(op.basename(tf)))
