==> processing/01c_determine_mriqc_outliers.ipynb <==
import os.path as op
from glob import glob

import numpy as np
import pandas as pd

dsets = ["dset-cambridge", "dset-camcan", "dset-cohen", "dset-dalenberg", "dset-dupre"]
metrics = [("gsr_x", "upper"), ("gsr_y", "upper"), ("fber", "lower")]

base_dir = "/home/data/nbc/misc-projects/Salo_PowerReplication/"
sub_dir = "derivatives/mriqc/"

for dset in dsets:
    print(dset)
    dset_bad_subs = []
    target_files = sorted(glob(op.join(base_dir, dset, sub_dir, "*_bold.tsv")))
    for i_file, target_file in enumerate(target_files):
        name = op.basename(target_file).split(".")[0]
        df = pd.read_table(target_file)
        
        if i_file == 0:
            other_files = sorted(glob(op.join(base_dir, dset, sub_dir, "*.tsv")))
            other_files = sorted(list(set(other_files) - set(target_files)))
            other_names = [op.basename(of).split(".")[0] for of in other_files]
            dset_df = pd.DataFrame(index=df["participant_id"], columns=other_names)
        
        dset_df[name] = np.nan

        all_bad_subs = []
        for metric, side in metrics:
            values = df[metric].values
            mean = values.mean()
            std = values.std()
            threshold = std * 3
            if side == "upper":
                max_val = mean + threshold
                bad_idx = df.loc[df[metric] > max_val].index
            elif side == "lower":
                min_val = mean - threshold
                bad_idx = df.loc[df[metric] < min_val].index
            bad_subs = df.loc[bad_idx, "participant_id"].tolist()
            all_bad_subs += bad_subs

        dset_df.loc[all_bad_subs, name] = 1
        dset_bad_subs += all_bad_subs
        print(f"\t{name}: {', '.join(all_bad_subs)}")
    dset_bad_subs = sorted(list(set(dset_bad_subs)))
    if dset == "dset-camcan":
        # Add one subject who fails fMRIPrep
        dset_bad_subs.append("sub-CC221585")
        dset_bad_subs = sorted(list(set(dset_bad_subs)))

    print(f"\n{dset}: {', '.join(dset_bad_subs)}\n")
    participants_file = op.join(base_dir, dset, "participants.tsv")
    participants_df = pd.read_table(participants_file)
    participants_df["exclude"] = 0
    participants_df.loc[participants_df["participant_id"].isin(dset_bad_subs), "exclude"] = 1
    participants_df.to_csv(participants_file, sep="\t", index=False)
